Added single multiple exe pact tests and changed the package structure
