+++ package.json
@@ -24,7 +24,7 @@
     "test-crossbrowser-e2e": "NODE_PATH=. codeceptjs run -c test/end-to-end/saucelabs.conf.js --steps",
     "test:smoke": "mocha test/smoke/healthcheck.js --timeout 25000 --reporter spec",
     "test-contract": "NODE_PATH=. LOG_LEVEL=error mocha 'test/contract/addresslookup/*.js' 'test/contract/generatepin/*.js' --timeout 15000 --reporter spec",
-    "test-pact": "NODE_PATH=. LOG_LEVEL=error mocha 'test/contract/pact/*.js' --timeout 15000 --reporter spec",
+    "test-pact": "NODE_PATH=. LOG_LEVEL=error mocha 'test/contract/pact/*/*.js' --timeout 15000 --reporter spec",
     "publish-pact": "NODE_PATH=. LOG_LEVEL=error mocha 'test/contract/publish/*.js' --timeout 15000 --reporter spec",
     "lint": "NODE_PATH=. eslint .",
     "eslint": "NODE_PATH=. eslint .",
