+++ test/contract/pact/intestacy/IntestacyFormDataClient.spec.js
@@ -5,51 +5,52 @@ const path = require('path');
 const chai = require('chai');
 const {Pact} = require('@pact-foundation/pact');
 const chaiAsPromised = require('chai-as-promised');
-const ProbateCoverSheetPdf = require('app/services/ProbateCoverSheetPdf');
+const IntestacyFormData = require('app/services/IntestacyFormData');
 const config = require('app/config');
-const assert = chai.assert;
 const getPort = require('get-port');
-const DOC_BODY_PAYLOAD = require('test/data/pacts/coverSheet');
+const expect = chai.expect;
+const FORM_DATA_BODY_PAYLOAD = require('test/data/pacts/intestacy/fromDataClient');
 
 chai.use(chaiAsPromised);
 
-describe('Pact ProbateCoverSheetPdf', () => {
+describe('Pact IntestacyFormData', () => {
 
     let MOCK_SERVER_PORT;
     let provider;
-    // (1) Create the Pact object to represent your provider
     getPort().then(portNumber => {
         MOCK_SERVER_PORT = portNumber;
         // (1) Create the Pact object to represent your provider
         provider = new Pact({
             consumer: 'probate_frontend',
-            provider: 'probate_orchestrator_service_documents_cover_sheet',
+            provider: 'probate_orchestrator_service_intestacy_forms',
             port: MOCK_SERVER_PORT,
-            log: path.resolve(process.cwd(), 'logs', 'pactProbateCoverSheetPdf.log'),
+            log: path.resolve(process.cwd(), 'logs', 'pactIntestacyFormDataClient.log'),
             dir: path.resolve(process.cwd(), config.services.pact.pactDirectory),
             logLevel: 'INFO',
             spec: 2
         });
     });
-
-    const req = {
+    const ctx = {
         sessionID: 'someSessionId',
         authToken: 'authToken',
         session: {
-            serviceAuthorization: 'someServiceAuthorization',
-            formdata: {
-                applicant: {address: 'addressLine1'},
-                ccdCase: {id: '123454'},
-                registry: {address: 'manchester'}
-            }
+            serviceAuthorization: 'someServiceAuthorization'
         }
     };
-
     function getRequestBody() {
-        const fullBody = {
-            bulkScanCoverSheet: DOC_BODY_PAYLOAD
-        };
+        const fullBody = JSON.parse(JSON.stringify(FORM_DATA_BODY_PAYLOAD));
+        fullBody.type = 'Intestacy';
         return fullBody;
+    }
+    function getExpectedResponseBody() {
+
+        const expectedJSON = JSON.parse(JSON.stringify(FORM_DATA_BODY_PAYLOAD));
+        expectedJSON.ccdCase = {
+            'id': 1535574519543819,
+            'state': 'Draft'
+        };
+        expectedJSON.type = 'Intestacy';
+        return expectedJSON;
     }
 
     // Setup a Mock Server before unit tests run.
@@ -60,55 +61,55 @@ describe('Pact ProbateCoverSheetPdf', () => {
     // if the calls are not seen.
     before(() =>
         provider.setup()
-    );
+    )
 
     // After each individual test (one or more interactions)
     // we validate that the correct request came through.
     // This ensures what we _expect_ from the provider, is actually
     // what we've asked for (and is what gets captured in the contract)
-    afterEach(() => provider.verify());
+    afterEach(() => provider.verify())
 
-    describe('when cover sheet doc is posted', () => {
-        describe('and is required to be downloaded', () => {
+    describe('when intestacy formdata is posted', () => {
+        describe('and is required to be persisted', () => {
             before(() => {
                 // (2) Start the mock server
                 provider.addInteraction({
                     // The 'state' field specifies a 'Provider State'
-                    state: 'probate_orchestrator_service generates cover sheet byte[] with success',
-                    uponReceiving: 'a request to POST cover sheet doc',
+                    state: 'probate_orchestrator_service persists intestacy formdata with success',
+                    uponReceiving: 'a request to POST intestacy formdata',
                     withRequest: {
                         method: 'POST',
-                        path: '/documents/generate/bulkScanCoversheet',
+                        path: '/forms/someemailaddress@host.com',
                         headers: {
-                            'Content-Type': 'application/businessdocument+json',
-                            'Session-Id': req.sessionID,
-                            'Authorization': req.authToken,
-                            'ServiceAuthorization': req.session.serviceAuthorization
+                            'Content-Type': 'application/json',
+                            'Session-Id': ctx.sessionID,
+                            'Authorization': ctx.authToken,
+                            'ServiceAuthorization': ctx.serviceAuthorization
                         },
                         body: getRequestBody()
                     },
                     willRespondWith: {
                         status: 200,
-                        headers: {'Content-Type': 'application/octet-stream'},
+                        headers: {'Content-Type': 'application/json'},
+                        body: getExpectedResponseBody()
                     }
-                });
-
-            });
+                })
+            })
 
             // (4) write your test(s)
             // Verify service client works as expected
             it('successfully validated form data', (done) => {
-                const coverSheetPdfClient = new ProbateCoverSheetPdf('http://localhost:' + MOCK_SERVER_PORT, req.sessionID);
-                const verificationPromise = coverSheetPdfClient.post(req);
-                assert.eventually.ok(verificationPromise).notify(done);
+                const formDataClient = new IntestacyFormData('http://localhost:' + MOCK_SERVER_PORT, 'someSessionId');
+                const verificationPromise = formDataClient.post('someemailaddress@host.com', FORM_DATA_BODY_PAYLOAD, ctx);
+                expect(verificationPromise).to.eventually.eql(getExpectedResponseBody()).notify(done);
             });
 
-            // (6) write the pact file for this consumer-provider pair,
-            // and shutdown the associated mock server.
-            // You should do this only _once_ per Provider you are testing.
+
         });
     });
-
+    // (6) write the pact file for this consumer-provider pair,
+    // and shutdown the associated mock server.
+    // You should do this only _once_ per Provider you are testing.
     after(() => {
         return provider.finalize();
     });
