+++ test/contract/pact/probate/ProbateFormDataClient.spec.js
@@ -5,31 +5,32 @@ const path = require('path');
 const chai = require('chai');
 const {Pact} = require('@pact-foundation/pact');
 const chaiAsPromised = require('chai-as-promised');
-const IntestacyFormData = require('app/services/IntestacyFormData');
+const ProbateFormData = require('app/services/ProbateFormData');
 const config = require('app/config');
-const getPort = require('get-port');
 const expect = chai.expect;
+const getPort = require('get-port');
+const FORM_DATA_BODY_PAYLOAD = require('test/data/pacts/probate/formDataClient');
 
 chai.use(chaiAsPromised);
 
-describe('Pact IntestacyFormData', () => {
+describe('Pact ProbateFormData', () => {
 
     let MOCK_SERVER_PORT;
     let provider;
+    // (1) Create the Pact object to represent your provider
     getPort().then(portNumber => {
         MOCK_SERVER_PORT = portNumber;
         // (1) Create the Pact object to represent your provider
         provider = new Pact({
             consumer: 'probate_frontend',
-            provider: 'probate_orchestrator_service_intestacy_forms',
+            provider: 'probate_orchestrator_service_probate_forms',
             port: MOCK_SERVER_PORT,
-            log: path.resolve(process.cwd(), 'logs', 'pactIntestacyFormDataClient.log'),
+            log: path.resolve(process.cwd(), 'logs', 'pactProbateFormData.log'),
             dir: path.resolve(process.cwd(), config.services.pact.pactDirectory),
             logLevel: 'INFO',
             spec: 2
         });
     });
-
     const ctx = {
         sessionID: 'someSessionId',
         authToken: 'authToken',
@@ -37,44 +38,9 @@ describe('Pact IntestacyFormData', () => {
             serviceAuthorization: 'someServiceAuthorization'
         }
     };
-
-    // Define expected payloads
-    const FORM_DATA_BODY_PAYLOAD =
-        {
-            'applicant': {
-                'email': 'someemailaddress@host.com',
-                'firstName': 'Jon',
-                'lastName': 'Snow',
-                'address': 'Pret a Manger St. Georges Hospital Blackshaw Road London SW17 0QT',
-                'postCode': 'SW17 0QT',
-                'phoneNumber': '123455678',
-                'addressFound': 'Yes',
-                'freeTextAddress': 'Pret a Manger St. Georges Hospital Blackshaw Road',
-                'adoptionInEnglandOrWales': 'Yes'
-            },
-            'deceased': {
-                'firstName': 'Ned',
-                'lastName': 'Stark',
-                'dob_date': '1930-01-01',
-                'dod_date': '2018-01-01',
-                'address': 'Winterfell, Westeros',
-                'addressFound': 'Yes',
-                'postCode': 'SW17 0QT',
-                'freeTextAddress': 'Winterfell, Westeros',
-                'alias': 'Yes',
-                'allDeceasedChildrenOverEighteen': 'Yes',
-                'anyDeceasedChildrenDieBeforeDeceased': 'No',
-                'anyDeceasedGrandchildrenUnderEighteen': 'No',
-                'anyChildren': 'No'
-            },
-            'declaration': {
-                'declarationCheckbox': 'Yes'
-            }
-        };
-
     function getRequestBody() {
         const fullBody = JSON.parse(JSON.stringify(FORM_DATA_BODY_PAYLOAD));
-        fullBody.type = 'Intestacy';
+        fullBody.type = 'PA';
         return fullBody;
     }
 
@@ -85,7 +51,7 @@ describe('Pact IntestacyFormData', () => {
             'id': 1535574519543819,
             'state': 'Draft'
         };
-        expectedJSON.type = 'Intestacy';
+        expectedJSON.type = 'PA';
         return expectedJSON;
     }
 
@@ -97,30 +63,68 @@ describe('Pact IntestacyFormData', () => {
     // if the calls are not seen.
     before(() =>
         provider.setup()
-    );
+    )
 
     // After each individual test (one or more interactions)
     // we validate that the correct request came through.
     // This ensures what we _expect_ from the provider, is actually
     // what we've asked for (and is what gets captured in the contract)
-    afterEach(() => provider.verify());
+    afterEach(() => provider.verify())
 
-    describe('when intestacy formdata is posted', () => {
+    context('when formdata is requested', () => {
+        describe('from a GET', () => {
+            before(() => {
+                // (2) Start the mock server
+                provider.addInteraction({
+                    // The 'state' field specifies a 'Provider State'
+                    state: 'probate_orchestrator_service gets formdata with success',
+                    uponReceiving: 'a request to GET probate formdata',
+                    withRequest: {
+                        method: 'GET',
+                        path: '/forms/someemailaddress@host.com',
+                        query: 'probateType=PA',
+                        headers: {
+                            'Content-Type': 'application/json',
+                            'Session-Id': ctx.sessionID,
+                            'Authorization': ctx.authToken,
+                            'ServiceAuthorization': ctx.serviceAuthorization
+                        }
+                    },
+                    willRespondWith: {
+                        status: 200,
+                        headers: {'Content-Type': 'application/json'},
+                        body: getExpectedResponseBody()
+                    }
+                });
+            });
+
+            // (4) write your test(s)
+            // Verify service client works as expected
+            it('successfully get form data', (done) => {
+                const formDataClient = new ProbateFormData('http://localhost:' + MOCK_SERVER_PORT, 'someSessionId');
+                const verificationPromise = formDataClient.get('someemailaddress@host.com', ctx.authToken, ctx.serviceAuthorization);
+                expect(verificationPromise).to.eventually.eql(getExpectedResponseBody()).notify(done);
+            });
+
+        });
+    });
+
+    context('when probate formdata is posted', () => {
         describe('and is required to be persisted', () => {
             before(() => {
                 // (2) Start the mock server
                 provider.addInteraction({
                     // The 'state' field specifies a 'Provider State'
-                    state: 'probate_orchestrator_service persists intestacy formdata with success',
-                    uponReceiving: 'a request to POST intestacy formdata',
+                    state: 'probate_orchestrator_service persists probate formdata with success',
+                    uponReceiving: 'a request to POST probate formdata',
                     withRequest: {
                         method: 'POST',
                         path: '/forms/someemailaddress@host.com',
                         headers: {
                             'Content-Type': 'application/json',
-                            'Session-Id': ctx.sessionID,
-                            'Authorization': ctx.authToken,
-                            'ServiceAuthorization': ctx.session.serviceAuthorization
+                            'Session-Id': 'someSessionId',
+                            'Authorization': 'authToken',
+                            'ServiceAuthorization': 'someServiceAuthorization'
                         },
                         body: getRequestBody()
                     },
@@ -135,7 +139,7 @@ describe('Pact IntestacyFormData', () => {
             // (4) write your test(s)
             // Verify service client works as expected
             it('successfully validated form data', (done) => {
-                const formDataClient = new IntestacyFormData('http://localhost:' + MOCK_SERVER_PORT, 'someSessionId');
+                const formDataClient = new ProbateFormData('http://localhost:' + MOCK_SERVER_PORT, 'someSessionId');
                 const verificationPromise = formDataClient.post('someemailaddress@host.com', FORM_DATA_BODY_PAYLOAD, ctx);
                 expect(verificationPromise).to.eventually.eql(getExpectedResponseBody()).notify(done);
             });
