+++ test/data/pacts/intestacy/submitDataClient.json
@@ -1,0 +1,65 @@
+{
+  "applicant": {
+    "email": "someemailaddress@host.com",
+    "firstName": "Jon",
+    "lastName": "Snow",
+    "address": "Pret a Manger St. Georges Hospital Blackshaw Road London SW17 0QT",
+    "postCode": "SW17 0QT",
+    "phoneNumber": "123455678",
+    "addressFound": "Yes",
+    "freeTextAddress": "Pret a Manger St. Georges Hospital Blackshaw Road",
+    "relationshipToDeceased": "adoptedChild",
+    "adoptionInEnglandOrWales": "Yes"
+  },
+  "deceased": {
+    "firstName": "Ned",
+    "lastName": "Stark",
+    "dob_date": "1930-01-01",
+    "dod_date": "2018-01-01",
+    "address": "Winterfell, Westeros",
+    "addressFound": "Yes",
+    "postCode": "SW17 0QT",
+    "freeTextAddress": "Winterfell, Westeros",
+    "alias": "Yes",
+    "otherNames": {
+      "name_0": {
+        "lastName": "North",
+        "firstName": "King"
+      }
+    },
+    "maritalStatus": "marriedCivilPartnership",
+    "divorcedInEnglandOrWales": "No",
+    "domiciledInEnglandOrWales": "Yes",
+    "spouseNotApplyingReason": "mentallyIncapable",
+    "otherChildren": "Yes",
+    "allDeceasedChildrenOverEighteen": "Yes",
+    "anyDeceasedChildrenDieBeforeDeceased": "No",
+    "anyDeceasedGrandchildrenUnderEighteen": "No",
+    "anyChildren": "No"
+  },
+  "iht": {
+    "form": "IHT205",
+    "method": "Through the HMRC online service",
+    "netValue": 10000.99,
+    "grossValue": 10000.99,
+    "identifier": "GOT123456"
+  },
+  "assets": {
+    "assetsOverseasNetValue": 100.99,
+    "assetsOverseas": "Yes"
+  },
+  "copies": {
+    "uk": 5,
+    "overseas": 6
+  },
+  "registry": {
+    "name": "Birmingham",
+    "email": "birmingham@email.com",
+    "address": "Line 1 Bham\nLine 2 Bham\nLine 3 Bham\nPostCode Bham",
+    "sequenceNumber": 20075
+  },
+  "declaration": {
+    "declarationCheckbox": "Yes"
+  },
+  "uploadDocumentUrl": "http://document-management/document/12345"
+}
