+++ test/data/pacts/probate/invalidNoTitleCheckAnswersSummary.json
@@ -1,8 +1,8 @@
 {
+  "pageTitle": "\n        Check your answers\n    ",
   "mainParagraph": "Check the information below carefully. This will form a record of your application for probate. It will also be stored as a public record, and will be able to be viewed online.",
   "sections": [
     {
-      "title": "\n        The will\n    ",
       "type": "heading-medium",
       "questionAndAnswers": [
         {
